services/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
ebooklib==0.18
jinja2==3.1.3
bleach==6.1.0
python-dateutil==2.9.0.post0
requests==2.31.0
python-multipart==0.0.9
